+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   discovered to be a directory lading will merge all .yaml files in that
   directory. Configuration files must be disjoint, else parsing will fail and
   lading will not start up.
-  
+
 ## Changed
 - Updated OpenTelemetry trace generator to generate more realistic traces: resource,
   scope, and attributes are populated, along with nested child spans.
+- The Datadog Intake has been changed from a blackhole to a target metrics
+  module.
 
 ## [0.29.2]
 ## Added